3921 client - components - Replace all instances of deprecated `ui/button` imports with updated `Button` component and refactor props accordingly
