+++ client/src/components/DatePicker/DatePicker.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import {Button} from '@/components/ui/button';
+import Button from '@/components/Button/Button';
 import {Calendar} from '@/components/ui/calendar';
 import {Popover, PopoverContent, PopoverTrigger} from '@/components/ui/popover';
 import {cn} from '@/shared/util/cn-utils';
