+++ client/src/components/MultiSelect/MultiSelect.tsx
@@ -1,5 +1,5 @@
 import Badge from '@/components/Badge/Badge';
-import {Button} from '@/components/ui/button';
+import Button from '@/components/Button/Button';
 import {
     Command,
     CommandEmpty,
